pyproject.toml
[tool.poetry]
name = "fastapi-langraph"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fastapi_langraph"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
langchain = "^0.2.6"
langchain-openai = "^0.1.9"
langgraph = "^0.1.1"
langchain-core = "^0.2.11"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"
pre-commit = "^3.7.0"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
